+++ CHANGELOG.md
@@ -317,4 +317,5 @@ Change main.c to use OUTPUTTYPE instead of float
 ### Added
 - Generic float DW Conv2D kernel and bindings.
 - Bias handling and computation for regular and DW Conv2D.
+- Empty bias handling for generic regular and DW Conv2D.
 - Tests for Conv2D regular and DW, with and without bias (and included them in the CI pipeline).